ofw_edge: Sample app adapted for Ofw
Simplified the Sample so it is applicable for Fyber's Offer Wall SDK.
Left only the MainScene which launches the Ofw directly.
We might want to bring back other utilities/possibility to request and then show but, for the time being a simpler approach seems better.
Used Unity 2018.4 for scene changes but  didn't commit the ProjectSettings.asset file.